Fixes for PR

+++ test/hypothesis_tests.jl
@@ -1,6 +1,6 @@
-h1 = History([1, 2, 3, 4], 0, 5)
-h_empty = History(Float64[], 0, 2)
-PP = PoissonProcess{Float32,NoMarks}
+h1 = History([1.0, 2.0, 3.0, 4.0], 0.0, 5.0)
+h_empty = History(Float64[], 0.0, 2.0)
+PP = PoissonProcess{Float64,NoMarks}
 pp = PoissonProcess()
 
 @testset "Statistics" begin
@@ -10,7 +10,7 @@ pp = PoissonProcess()
     @test statistic(KSDistance{Exponential}, pp, h_empty) ≈ 1
 end
 
-h2 = History(collect(0:999), 0, 1000)
+h2 = History(collect(0.0:999.0), 0.0, 1000.0)
 
 @testset "BootstrapTest" begin
     @test_throws ArgumentError BootstrapTest(KSDistance{Uniform}, PP, h_empty)
